requirements.txt
opencv-python==4.8.0.76
moviepy==1.0.3
Pillow==10.0.0